promotion form, get header, get banner

+++ VZ-NodeJs-ThaiVietjet-Promotion/src/services/bannerService.js
@@ -22,14 +22,24 @@ let createBanner = (data) => {
     })
 }
 
-let getAllBanner = () => {
+let decodeImageOfBanner = (banner) => {
+    if (banner.imageMobile) banner.imageMobile = func.DECODE_IMAGE(banner.imageMobile)
+    if (banner.imageDesktop) banner.imageDesktop = func.DECODE_IMAGE(banner.imageDesktop)
+}
+
+let getBanner = (id) => {
     return new Promise(async (resolve, reject) => {
         try {
-            let data = await db.Banner.findAll()
-            data && data.map((item) => {
-                if (item.imageMobile) item.imageMobile = new Buffer(item.imageMobile, 'base64').toString('binary')
-                if (item.imageDesktop) item.imageDesktop = new Buffer(item.imageDesktop, 'base64').toString('binary')
-            })
+            let data
+            if (id) {
+                data = await db.Banner.findOne({ where: { id: id } })
+                decodeImageOfBanner(data)
+            } else {
+                data = await db.Banner.findAll()
+                data && data.map((item) => {
+                    decodeImageOfBanner(item)
+                })
+            }
             resolve(resolveObj.GET(data))
         } catch (e) {
             reject(e)
@@ -82,8 +92,12 @@ let deleteBanner = (data) => {
                 resolve(resolveObj.MISSING_PARAMETERS)
                 return
             }
-
             await db.sequelize.transaction(async (t) => {
+                let campaign = await db.Campaign.findOne({ where: { bannerId: data.id } })
+                if (campaign) {
+                    resolve(resolveObj.EXIST_REF_KEY)
+                    throw new Error()
+                }
                 await db.Banner.destroy({ where: { id: data.id }, transaction: t })
             })
             resolve(resolveObj.DELETE_SUCCEED())
@@ -95,7 +109,7 @@ let deleteBanner = (data) => {
 
 module.exports = {
     createBanner,
-    getAllBanner,
+    getBanner,
     getBannerById,
     updateBanner,
     deleteBanner,
